pyproject.toml
[project]
name = "fortress-eval"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "vllm",
    "openai",
    "datasets",
    "transformers",
    "pyyaml",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
